repos/zkml-vector_transform/src/vector_transform.py

import numpy as np
from ezkl import Circuit, Verifier, Prover

class VectorTransform:
    def __init__(self, m, n):
        self.m = m  # Number of input vectors
        self.n = n  # Dimension of vectors

    def transform(self, vectors):
        """
        Perform the many-to-one vector transformation.
        """
        if len(vectors) != self.m or any(len(v) != self.n for v in vectors):
            raise ValueError("Invalid input dimensions")
        
        # Simple example: sum all vectors
        result = np.sum(vectors, axis=0)
        return result

    def generate_proof(self, vectors, result):
        """
        Generate a ZK proof for the vector transformation.
        """
        circuit = Circuit()
        
        # Define inputs
        inputs = circuit.input("vectors", shape=(self.m, self.n))
        
        # Define the computation
        computed_result = circuit.sum(inputs, axis=0)
        
        # Define the output
        circuit.output(computed_result)
        
        # Create a prover
        prover = Prover(circuit)
        
        # Generate the proof
        proof = prover.prove({
            "vectors": vectors,
            "output": result
        })
        
        return proof

    def verify_proof(self, proof, public_inputs):
        """
        Verify the ZK proof of the vector transformation.
        """
        verifier = Verifier(self.circuit)
        return verifier.verify(proof, public_inputs)

